% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedgui.R
\name{impliedGUI}
\alias{impliedGUI}
\title{Run the implied GUI}
\usage{
impliedGUI(...)
}
\arguments{
\item{...}{Passed on to the \code{\link[shiny:runApp]{shiny::runApp}} function.}
}
\description{
Run this function to start the grapichal user interface (GUI) for the implied package. Please refer to the documentation and reference in the implied package for more information.
}
\note{
This function is based on code from Dean Attali's blog (https://deanattali.com/2015/04/21/r-package-shiny-app/)
}
